requirements.txt
flask>=3.0
websockets>=12.0
pyyaml>=6.0
pyserial>=3.5
